update component blog
