+++ app/page.tsx
@@ -1,23 +1,7 @@
-import { fetchGraphQL } from "@/api";
-import { Blog, AllBlogResponse, AllBlogQuery } from "@/interfaces/blog";
+import getAllBlog from "@/api/queries/getBlog";
 
 export default async function Home() {
-  const results: AllBlogResponse = (await fetchGraphQL(
-    AllBlogQuery
-  )) as AllBlogResponse;
-
-  const posts: Partial<Blog>[] = [];
-
-  results.data.allBlog.results.forEach((post: Partial<Blog>) => {
-    posts.push({
-      id: post.id,
-      description: post.description,
-      name: post.name,
-      title: post.title,
-      publishDate: post.publishDate,
-      blogImage: post.blogImage,
-    });
-  });
+  const posts = await getAllBlog();
 
   return (
     <main>
